Makefile: Disable "format-truncation" warning globally
This was introduced in GCC 7.x: https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html#index-Wformat-truncation

This warning doesn't really make sense in the context of the kernel
because there is no way for the values it complains about to be cut
off. For example, it complains about a cpu int getting truncated,
when in reality it will never be over 7 for this device (octacore).

 Conflicts:
	Makefile

Change-Id: I4f22cd225545841c3ec57203a9ef38c701da5e8f
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Albert I <krascgq@outlook.co.id>
Signed-off-by: Ad!thya R <gh0strider.2k18.reborn@gmail.com>